Add more methods in FriendRequestDao

+++ puluo_app_backend/core/src/main/java/com/puluo/api/social/DeleteFriendAPI.java
@@ -9,6 +9,7 @@ import com.puluo.dao.PuluoDSI;
 import com.puluo.dao.PuluoUserFriendshipDao;
 import com.puluo.dao.impl.DaoApi;
 import com.puluo.dao.impl.PuluoFriendRequestDaoImpl;
+import com.puluo.entity.FriendRequestStatus;
 import com.puluo.entity.PuluoFriendRequest;
 import com.puluo.entity.PuluoUserFriendship;
 import com.puluo.entity.impl.PuluoFriendInfo;
@@ -35,9 +36,13 @@ public class DeleteFriendAPI extends PuluoAPI<PuluoDSI,DeleteFriendResult> {
 		log.info(String.format("用户:%s和用户:%s解除好友关系",fromUser,toUser));
 		PuluoUserFriendshipDao friendshipdao = dsi.friendshipDao();
 		PuluoFriendRequestDaoImpl friendRequestDao = (PuluoFriendRequestDaoImpl)dsi.friendRequestDao();
-		PuluoFriendRequest request = friendRequestDao.getFriendRequestByUsers(fromUser,toUser);
+		PuluoFriendRequest request = friendRequestDao.getFriendRequestByUsers(fromUser,toUser,FriendRequestStatus.Approved);
 		friendshipdao.deleteOneFriend(fromUser,toUser);
-		friendRequestDao.deleteByReqUUID(request.requestUUID());
+		if(request!=null){
+			friendRequestDao.deleteByReqUUID(request.requestUUID());
+		}else{
+			log.warn(String.format("用户%s和%s之间的好友请求不存在", fromUser,toUser));
+		}
 		List<String> newFriends = new ArrayList<String>();
 		PuluoUserFriendship fr = friendshipdao.getFriendListByUUID(fromUser);
 		for(PuluoFriendInfo info:fr.friends()){
